Fix grpc e2e tests
