+++ rln-prover/prover/src/karma_sc_listener.rs
@@ -359,7 +359,7 @@ mod tests {
             tree_depth: MERKLE_TREE_HEIGHT,
         };
 
-        let db_conn = create_database_connection_1(file!(), function_name!())
+        let (_, db_conn) = create_database_connection_1(file!(), function_name!())
             .await
             .unwrap();
         let user_db_service = UserDbService::new(
